requirements.txt


aiohttp                  == 3.8.1
aiosignal                == 1.2.0
altair                   == 4.1.0
argon2-cffi              == 21.1.0
astor                    == 0.8.1
async-timeout            == 4.0.1
attrs                    == 21.2.0
backcall                 == 0.2.0
base58                   == 2.1.1
bleach                   == 4.1.0
blinker                  == 1.4
cachetools               == 4.2.4
certifi                  == 2021.10.8
cffi                     == 1.15.0
charset-normalizer       == 2.0.7
click                    == 7.1.2
click-plugins            == 1.1.1
colorama                 == 0.4.4
cycler                   == 0.11.0
DateTime                 == 4.3
debugpy                  == 1.5.1
decorator                == 5.1.0
defusedxml               == 0.7.1
docutils                 == 0.18
dragonfly-core           == 1.30.2
dragonfly-energy         == 1.15.10
dragonfly-schema         == 1.6.57
entrypoints              == 0.3
et-xmlfile               == 1.1.0
fonttools                == 4.28.1
frozenlist               == 1.2.0
gitdb                    == 4.0.9
GitPython                == 3.1.24
greenlet                 == 1.1.2
honeybee-core            == 1.48.5
honeybee-energy          == 1.83.13
honeybee-energy-standards == 2.2.1
honeybee-radiance         ==1.49.13
honeybee-radiance-command ==1.16.0
honeybee-radiance-folder  ==2.7.0
honeybee-schema           ==1.47.0
honeybee-standards        ==2.0.5
honeybee-vtk              ==0.30.7
idna                      ==3.3
importlib-metadata        ==4.8.2
ipykernel                 ==6.5.1
ipython                   ==7.30.0
ipython-genutils          ==0.2.0
ipywidgets                ==7.6.5
jedi                      ==0.18.1
Jinja2                    ==3.0.3
jsonschema                ==3.2.0
jupyter-client            ==7.1.0
jupyter-core              ==4.9.1
jupyterlab-pygments       ==0.1.2
jupyterlab-widgets        ==1.0.2
kiwisolver                ==1.3.2
ladybug-comfort           ==0.12.35
ladybug-core              ==0.39.26
ladybug-geometry          ==1.23.19
ladybug-geometry-polyskel ==1.3.40
lbt-honeybee              ==0.5.177
lbt-ladybug               ==0.25.97
lockfile                  ==0.12.2
luigi                     ==3.0.3
MarkupSafe                ==2.0.1
matplotlib                ==3.5.0
matplotlib-inline         ==0.1.3
mistune                   ==0.8.4
multidict                 ==5.2.0
nbclient                  ==0.5.9
nbconvert                 ==6.3.0
nbformat                  ==5.1.3
nest-asyncio              ==1.5.1
notebook                  ==6.4.6
numpy                     ==1.21.4
openpyxl                  ==3.0.9
packaging                 ==21.2
pandas                    ==1.3.4
pandocfilters             ==1.5.0
parso                     ==0.8.3
pickleshare               ==0.7.5
Pillow                    ==8.4.0
pip                       ==21.3.1
pollination-dsl           ==0.12.5
pollination-sdk           ==0.10.17
pollination-streamlit     ==0.1.1
prometheus-client         ==0.12.0
prompt-toolkit            ==3.0.23
protobuf                  ==3.19.1
pyarrow                   ==6.0.1
pycparser                 ==2.21
pydantic                  ==1.8.2
pydantic-openapi-helper   ==0.2.7
pydeck                    ==0.7.1
Pygments                  ==2.10.0
Pympler                   ==0.9
pyparsing                 ==2.4.7
pyrsistent                ==0.18.0
PySimpleGUI               ==4.55.1
python-daemon             ==2.3.0
python-dateutil           ==2.8.2
pytz                      ==2021.3
pytz-deprecation-shim     ==0.1.0.post0
pywin32                   ==302
pywinpty                  ==1.1.6
PyYAML                    ==5.4.1
pyzmq                     ==22.3.0
queenbee                  ==1.26.2
queenbee-local            ==0.3.16
queenbee-pollination      ==0.7.5
requests                  ==2.26.0
Send2Trash                ==1.8.0
setuptools                ==56.0.0
setuptools-scm            ==6.3.2
six                       ==1.16.0
smmap                     ==5.0.0
SQLAlchemy                ==1.4.27
streamlit                 ==1.2.0
streamlit-vtkjs           ==0.0.7a0
tabulate                  ==0.8.9
tenacity                  ==6.3.1
terminado                 ==0.12.1
testpath                  ==0.5.0
toml                      ==0.10.2
tomli                     ==1.2.2
toolz                     ==0.11.2
tornado                   ==6.1
traitlets                 ==5.1.1
typing_extensions         ==4.0.0
tzdata                    ==2021.5
tzlocal                   ==4.1
urllib3                   ==1.26.7
validators                ==0.18.2
vtk                       ==9.1.0
watchdog                  ==2.1.6
wcwidth                   ==0.2.5
webencodings              ==0.5.1
wheel                     ==0.37.0
widgetsnbextension        ==3.5.2
wslink                    ==1.1.1
yarl                      ==1.7.2
zipp                      ==3.6.0
zope.interface            ==5.4.0